UIIN-1920 Move some utility to utils
